refactor(roles): hoist slugify into lib/utils
The transliteration table lived inside the add-role dialog; onboarding needs
the same helper.
