+++ src/lib/utils.ts
@@ -20,6 +20,56 @@ export function formatCurrency(value: number) {
   }).format(value)
 }
 
+const TRANSLIT: Record<string, string> = {
+  а: 'a',
+  б: 'b',
+  в: 'v',
+  г: 'g',
+  д: 'd',
+  е: 'e',
+  ё: 'e',
+  ж: 'zh',
+  з: 'z',
+  и: 'i',
+  й: 'y',
+  к: 'k',
+  л: 'l',
+  м: 'm',
+  н: 'n',
+  о: 'o',
+  п: 'p',
+  р: 'r',
+  с: 's',
+  т: 't',
+  у: 'u',
+  ф: 'f',
+  х: 'h',
+  ц: 'ts',
+  ч: 'ch',
+  ш: 'sh',
+  щ: 'sch',
+  ъ: '',
+  ы: 'y',
+  ь: '',
+  э: 'e',
+  ю: 'yu',
+  я: 'ya',
+}
+
+/** Название → slug: транслит кириллицы, латиница/цифры, дефисы. */
+export function slugify(input: string): string {
+  let out = ''
+  for (const ch of input.toLowerCase()) {
+    if (ch in TRANSLIT) out += TRANSLIT[ch]
+    else if (/[a-z0-9]/.test(ch)) out += ch
+    else out += '-'
+  }
+  return out
+    .replace(/-+/g, '-')
+    .replace(/^[^a-z]+/, '')
+    .replace(/-+$/, '')
+}
+
 export const DaysOfWeek = {
   short: ['Вс', 'Пн', 'Вт', 'Ср', 'Чт', 'Пт', 'Сб'],
   full: ['Воскресенье', 'Понедельник', 'Вторник', 'Среда', 'Четверг', 'Пятница', 'Суббота'],
